+++ src/include/NovaC.hpp
@@ -6,3 +6,5 @@
 #include "novac/assets/atomic/AtomicController.hpp"
 #include "novac/assets/essentials/EssentialsController.hpp"
 #include "novac/assets/types/TypeController.hpp"
+#include "novac/assets/types/LayoutController.hpp"
+#include "novac/assets/types/aggregate/StructType.hpp"